PTY.jl

Documentation for PTY.jl

PTY.TR for Terraria

PTY.TR.CombGate — Type

(AndGate|XorGate){T<:Integer, V<:AbstractVector{<:Integer}}(gatestate::Bool, args::T, lampstates::V) <:AbstractGate
CombGate{T, V} = Union{AndGate{T, V}, XorGate{T, V}}

Combinatory gates in Terraria. gatestate tells the default (when all inputs are zero) state of the gate. args tells the number of inputs. lampstates lists how the inputs are wired to each logic lamp.

Example

Consider AndGate(false, 4, [0x12, 0x09, 0x04]). It's an AND gate whose default state is OFF. Since there are 4 args (named a-d), we consider the last 5 bits of each entry of lampstates. The leading bit tells the default state of the lamps. The other 4 bits (from lower to higher) tell which args are wired to the lamp. i.e.,

0x12 = 0b10010 means the default state of the first lamp is ON and only b is connected to the lamp.

0x09 = 0b01001 means the default state of the second lamp is OFF and a and d are connected to the lamp.

0x04 = 0b00100 means the default state of the third lamp is OFF and only c is connected to the lamp.

julia> TR.AndGate(false, 4, [0x12, 0x09, 0x04])
&(~b, ad, c)

The output is visualized via gate2string.

PTY.TR.AndGate — Type

(AndGate|XorGate){T<:Integer, V<:AbstractVector{<:Integer}}(gatestate::Bool, args::T, lampstates::V) <:AbstractGate
CombGate{T, V} = Union{AndGate{T, V}, XorGate{T, V}}

Combinatory gates in Terraria. gatestate tells the default (when all inputs are zero) state of the gate. args tells the number of inputs. lampstates lists how the inputs are wired to each logic lamp.

Example

Consider AndGate(false, 4, [0x12, 0x09, 0x04]). It's an AND gate whose default state is OFF. Since there are 4 args (named a-d), we consider the last 5 bits of each entry of lampstates. The leading bit tells the default state of the lamps. The other 4 bits (from lower to higher) tell which args are wired to the lamp. i.e.,

0x12 = 0b10010 means the default state of the first lamp is ON and only b is connected to the lamp.

0x09 = 0b01001 means the default state of the second lamp is OFF and a and d are connected to the lamp.

0x04 = 0b00100 means the default state of the third lamp is OFF and only c is connected to the lamp.

julia> TR.AndGate(false, 4, [0x12, 0x09, 0x04])
&(~b, ad, c)

The output is visualized via gate2string.

PTY.TR.XorGate — Type

(AndGate|XorGate){T<:Integer, V<:AbstractVector{<:Integer}}(gatestate::Bool, args::T, lampstates::V) <:AbstractGate
CombGate{T, V} = Union{AndGate{T, V}, XorGate{T, V}}

Combinatory gates in Terraria. gatestate tells the default (when all inputs are zero) state of the gate. args tells the number of inputs. lampstates lists how the inputs are wired to each logic lamp.

Example

Consider AndGate(false, 4, [0x12, 0x09, 0x04]). It's an AND gate whose default state is OFF. Since there are 4 args (named a-d), we consider the last 5 bits of each entry of lampstates. The leading bit tells the default state of the lamps. The other 4 bits (from lower to higher) tell which args are wired to the lamp. i.e.,

0x12 = 0b10010 means the default state of the first lamp is ON and only b is connected to the lamp.

0x09 = 0b01001 means the default state of the second lamp is OFF and a and d are connected to the lamp.

0x04 = 0b00100 means the default state of the third lamp is OFF and only c is connected to the lamp.

julia> TR.AndGate(false, 4, [0x12, 0x09, 0x04])
&(~b, ad, c)

The output is visualized via gate2string.

PTY.TR.CombLogic — Method

CombLogic(lamps, inputs, outputs::AbstractVector{Bool})

Return the lists of possible combinatory logics for given inputs and outputs.

Arguments

lamps::Integer: the number of logic lamps on the gate, from 2 to 7.

inputs::AbstractVector{<:Integer}: the list of input values specified by the lower bits of entries, with a 1 in front of the first bit.

outputs::AbstractVector{Bool}: the expected return values of the logic. This should have the same length as inputs, although it's not strictly checked.

Example

Imagine we want to check if a 4-bit input can be divided by 3. Further, the input is expected to have only one decimal digit, i.e. in 0:9. We don't care how the logic behaves for inputs 10:15. The truth table would be

in	out		in	out

0000	true		0101	false

0001	false		0110	true

0010	false		0111	false

0011	true		1000	false

0100	false		1001	true

Now we try to find a configuration with least lamps.

julia> inputs = 16:25; # `0:9` with the 5th last bit set to `1`

julia> outputs = [true, false, false, true, false, false, true, false, false, true];

julia> TR.CombLogic(2, inputs, outputs) # there is no 2-lamp logic
PTY.TR.CombGate{Int64, StaticArraysCore.SVector{2, UInt8}}[]

julia> TR.CombLogic(3, inputs, outputs) # there are 12 3-lamp logics, all of which use XOR gates
12-element Vector{PTY.TR.CombGate{Int64, StaticArraysCore.SVector{3, UInt8}}}:
 ^(~b, ad, c)
 ^(~b, abd, bc)
 ^(~ac, ad, ab)
 ^(~ac, abd, a)
 ^(~ac, cd, bc)
 ^(~ac, bcd, c)
 ^(~bd, c, a)
 ^(~bd, bc, ab)
 ^(~bd, cd, ad)
 ^(~bd, bcd, abd)
 ^(~acd, bc, c)
 ^(~acd, abd, ad)

See also CombGate

PTY.TR.gate2string — Method

gate2string(G::CombGate)

Convert a CombGate into a readable string. See also CombGate, lampstate2string, logicsymb.

PTY.TR.lampstate2string — Method

lampstate2string(lampstate::Integer, args::Integer)

Convert binary coded lampstate into a readable string. args tells the number of distinct inputs.

Examples

julia> TR.lampstate2string(0b1111, 4)
"abcd"

julia> TR.lampstate2string(0b11101, 4)
"~acd"

julia> TR.lampstate2string(0b10000, 4)
"~"

PTY.TR.logicsymb — Method

logicsymb(::typeof(AND|XOR)|::CombGate)

Return '&' or '^' according to input.

PTY.SpecFun for Special functions

PTY.SpecFun.fracpochhammer — Method

fracpochhammer(a, b, stepa, stepb, n)

Similar to fracpochhammer(a, b, n), except that the steps of the Pochhammer symbols are not necessarily $1$.

Examples

julia> SpecFun.fracpochhammer(1, 2, 0.5, 1, 3) # (1 * 1.5 * 2) / (2 * 3 * 4)
0.125

PTY.SpecFun.fracpochhammer — Method

fracpochhammer(a, b, n)

Calculate the fraction of two Pochhammer symbols $\frac{(a)_n}{(b)_n}$ by multiplying the fractions. This approach reduces the risk of overflow/underflow when $n$ is large.

Examples

julia> SpecFun.fracpochhammer(1, 2, 3) # (1 * 2 * 3) / (2 * 3 * 4)
0.25

PTY.SpecFun.mittleff — Method

mittleff(α, β, γ, z)
mittleff(α, β, z) = mittleff(α, β, 1, z)
mittleff(α, z) = mittleff(α, 1, z)

Compute the Mittag-Leffler function of one variable z. The codes are translated from https://uk.mathworks.com/matlabcentral/fileexchange/48154-the-mittag-leffler-function