PTY.jl

Documentation for PTY.jl

Type piracies by PTY.Helper

// falls back to / for float inputs

: for range construction now supports complex arithmetic progressions

AbstractArray{T, N}(A::BroadcastArray{S, N}) for eltype conversion

eps supports complex inputs

ceil and floor support complex inputs.

PTY.TR for Terraria

See also $\\mathbb{Z}_2$-linear-algebra.

PTY.TR.CombGate — Type

(AndGate|XorGate){T<:Integer, V<:AbstractVector{<:Integer}}(gatestate::Bool, args::T, lampstates::V) <:AbstractGate
CombGate{T, V} = Union{AndGate{T, V}, XorGate{T, V}}

Combinatory gates in Terraria. gatestate tells the default (when all inputs are zero) state of the gate. args tells the number of inputs. lampstates lists how the inputs are wired to each logic lamp.

Example

Consider AndGate(false, 4, [0x12, 0x09, 0x04]). It's an AND gate whose default state is OFF. Since there are 4 args (named a-d), we consider the last 5 bits of each entry of lampstates. The leading bit tells the default state of the lamps. The other 4 bits (from lower to higher) tell which args are wired to the lamp. i.e.,

0x12 = 0b10010 means the default state of the first lamp is ON and only b is connected to the lamp.

0x09 = 0b01001 means the default state of the second lamp is OFF and a and d are connected to the lamp.

0x04 = 0b00100 means the default state of the third lamp is OFF and only c is connected to the lamp.

julia> TR.AndGate(false, 4, [0x12, 0x09, 0x04])
&(~b, ad, c)

The output is visualized via gate2string.

PTY.TR.AndGate — Type

(AndGate|XorGate){T<:Integer, V<:AbstractVector{<:Integer}}(gatestate::Bool, args::T, lampstates::V) <:AbstractGate
CombGate{T, V} = Union{AndGate{T, V}, XorGate{T, V}}

Combinatory gates in Terraria. gatestate tells the default (when all inputs are zero) state of the gate. args tells the number of inputs. lampstates lists how the inputs are wired to each logic lamp.

Example

Consider AndGate(false, 4, [0x12, 0x09, 0x04]). It's an AND gate whose default state is OFF. Since there are 4 args (named a-d), we consider the last 5 bits of each entry of lampstates. The leading bit tells the default state of the lamps. The other 4 bits (from lower to higher) tell which args are wired to the lamp. i.e.,

0x12 = 0b10010 means the default state of the first lamp is ON and only b is connected to the lamp.

0x09 = 0b01001 means the default state of the second lamp is OFF and a and d are connected to the lamp.

0x04 = 0b00100 means the default state of the third lamp is OFF and only c is connected to the lamp.

julia> TR.AndGate(false, 4, [0x12, 0x09, 0x04])
&(~b, ad, c)

The output is visualized via gate2string.

PTY.TR.XorGate — Type

(AndGate|XorGate){T<:Integer, V<:AbstractVector{<:Integer}}(gatestate::Bool, args::T, lampstates::V) <:AbstractGate
CombGate{T, V} = Union{AndGate{T, V}, XorGate{T, V}}

Combinatory gates in Terraria. gatestate tells the default (when all inputs are zero) state of the gate. args tells the number of inputs. lampstates lists how the inputs are wired to each logic lamp.

Example

Consider AndGate(false, 4, [0x12, 0x09, 0x04]). It's an AND gate whose default state is OFF. Since there are 4 args (named a-d), we consider the last 5 bits of each entry of lampstates. The leading bit tells the default state of the lamps. The other 4 bits (from lower to higher) tell which args are wired to the lamp. i.e.,

0x12 = 0b10010 means the default state of the first lamp is ON and only b is connected to the lamp.

0x09 = 0b01001 means the default state of the second lamp is OFF and a and d are connected to the lamp.

0x04 = 0b00100 means the default state of the third lamp is OFF and only c is connected to the lamp.

julia> TR.AndGate(false, 4, [0x12, 0x09, 0x04])
&(~b, ad, c)

The output is visualized via gate2string.

PTY.TR.Z2ColMat — Type

Z2ColMat{C<:Integer, R<:Integer}(data::Vector{C}, size::Int)
Z2ColMat(undef, m, n)
Z2ColMat(::AbstractMatrix)

$\mathbb{Z}_2$ matrix stored in a Vector where each entry represents a column. size specifies the height. C and R are used for dispatch; they specify the integer type used to represent a column vector and a row vector respectively. See also bit2type, Z2RowMat

Example

julia> TR.Z2ColMat([1, 3, 5], 3)
3×3 PTY.TR.Z2ColMat{UInt8, UInt8}:
 1  1  1
 0  1  0
 0  0  1

PTY.TR.Z2RowMat — Type

Z2RowMat{C<:Integer, R<:Integer}(data::Vector{R}, size::Int)
Z2RowMat(undef, m, n)
Z2RowMat(::AbstractMatrix)

$\mathbb{Z}_2$ matrix stored in a Vector where each entry represents a row. size specifies the width. C and R are used for dispatch; they specify the integer type used to represent a column vector and a row vector respectively. See also bit2type, Z2ColMat

Example

julia> TR.Z2RowMat([1, 3, 5], 3)
3×3 PTY.TR.Z2RowMat{UInt8, UInt8}:
 1  0  0
 1  1  0
 1  0  1

PTY.TR.Z2Vector — Type

Z2Vector{T<:Integer}(data::T, size::Int)
Z2Vector(::AbstractVector)

$\mathbb{Z}_2$ vector stored in an integer data where each bit represents an entry. size specifies the length.

Example

julia> TR.Z2Vector(0b10110, 5)
5-element PTY.TR.Z2Vector{UInt8}:
 0
 1
 1
 0
 1

PTY.TR.CombLogic — Method

CombLogic(lamps, inputs, outputs::AbstractVector{Bool})

Return the lists of possible combinatory logics for given inputs and outputs.

Arguments

lamps::Integer: the number of logic lamps on the gate, from 2 to 7.

inputs::AbstractVector{<:Integer}: the list of input values specified by the lower bits of entries, with a 1 in front of the first bit.

outputs::AbstractVector{Bool}: the expected return values of the logic. This should have the same length as inputs, although it's not strictly checked.

Example

Imagine we want to check if a 4-bit input can be divided by 3. Further, the input is expected to have only one decimal digit, i.e. in 0:9. We don't care how the logic behaves for inputs 10:15. The truth table would be

in	out		in	out

0000	true		0101	false

0001	false		0110	true

0010	false		0111	false

0011	true		1000	false

0100	false		1001	true

Now we try to find a configuration with least lamps.

julia> inputs = 16:25; # `0:9` with the 5th last bit set to `1`

julia> outputs = [true, false, false, true, false, false, true, false, false, true];

julia> TR.CombLogic(2, inputs, outputs) # there is no 2-lamp logic
PTY.TR.CombGate{Int64, StaticArraysCore.SVector{2, UInt8}}[]

julia> TR.CombLogic(3, inputs, outputs) # there are 12 3-lamp logics, all of which use XOR gates
12-element Vector{PTY.TR.CombGate{Int64, StaticArraysCore.SVector{3, UInt8}}}:
 ^(~b, ad, c)
 ^(~b, abd, bc)
 ^(~ac, ad, ab)
 ^(~ac, abd, a)
 ^(~ac, cd, bc)
 ^(~ac, bcd, c)
 ^(~bd, c, a)
 ^(~bd, bc, ab)
 ^(~bd, cd, ad)
 ^(~bd, bcd, abd)
 ^(~acd, bc, c)
 ^(~acd, abd, ad)

See also CombGate

PTY.TR.bit2type — Method

bit2type(n::Integer) :: DataType

Returns minimal integer type that contains n bits

Examples

julia> TR.bit2type(5)
UInt8

julia> TR.bit2type(9)
UInt16

julia> TR.bit2type(32)
UInt32

julia> TR.bit2type(33)
UInt64

julia> TR.bit2type(100)
BigInt

PTY.TR.gate2string — Method

gate2string(G::CombGate)

Convert a CombGate into a readable string. See also CombGate, lampstate2string, logicsymb.

PTY.TR.getbit — Method

getbit(x::Integer, n::Integer)

Get the n-th (counting from zero) last bit of x. See also setbit

Example

julia> [TR.getbit(0b10110, n) for n in 0:4]
5-element Vector{Bool}:
 0
 1
 1
 0
 1

PTY.TR.lampstate2string — Method

lampstate2string(lampstate::Integer, args::Integer)

Convert binary coded lampstate into a readable string. args tells the number of distinct inputs.

Examples

julia> TR.lampstate2string(0b1111, 4)
"abcd"

julia> TR.lampstate2string(0b11101, 4)
"~acd"

julia> TR.lampstate2string(0b10000, 4)
"~"

PTY.TR.logicsymb — Method

logicsymb(::typeof(AND|XOR)|::CombGate)

Return '&' or '^' according to input.

PTY.TR.setbit — Method

setbit(x::Integer, y::Bool, n::Integer)

Set the n-th (counting from zero) last bit of x to y. See also getbit.

Example

julia> string(TR.setbit(0b10110, true, 0), base = 2)
"10111"

julia> string(TR.setbit(0b10110, false, 0), base = 2)
"10110"

PTY.TR.type2bit — Method

type2bit(::Type{T})

The inverse of bit2type.

PTY.TR.z2number — Method

z2number(::Number)
z2number(::AbstractMatrix{<:Number})

Convert Number to Bool in the $\mathbb{Z}_2$ sense.

Examples

julia> TR.z2number(25)
true

julia> TR.z2number(1:5)
5-element BitVector:
 1
 0
 1
 0
 1

PTY.SpecFun for special functions

PTY.SpecFun.fracpochhammer — Method

fracpochhammer(a, b, n)

Calculate the fraction of two Pochhammer symbols $\frac{(a)_n}{(b)_n}$ by multiplying the fractions. This approach reduces the risk of overflow/underflow when $n$ is large.

Examples

julia> SpecFun.fracpochhammer(1, 2, 3) # (1 * 2 * 3) / (2 * 3 * 4)
0.25

fracpochhammer(a, b, stepa, stepb, n)

Similar to fracpochhammer(a, b, n), except that the steps of the Pochhammer symbols are not necessarily $1$.

Examples

julia> SpecFun.fracpochhammer(1, 2, 0.5, 1, 3) # (1 * 1.5 * 2) / (2 * 3 * 4)
0.125

PTY.SpecFun.mittleff — Method

mittleff(α, β, γ, z)
mittleff(α, β, z) = mittleff(α, β, 1, z)
mittleff(α, z) = mittleff(α, 1, z)

Compute the Mittag-Leffler function of one variable z. The codes are translated from https://uk.mathworks.com/matlabcentral/fileexchange/48154-the-mittag-leffler-function

PTY.ContFrac for continued fractions

PTY.ContFrac.ContinuedFraction — Type

ContinuedFraction(b0, a, b)

The continued fraction $b_0+\frac{a_1}{b_1+\frac{a_2}{b_2+\frac{a_3}{b_3+\frac{a_4}{b_4+\cdots}}}}$. Each entry represents a truncated fraction.

Example

julia> using InfiniteArrays, FillArrays

julia> ContFrac.ContinuedFraction(1, Ones{Int}(∞), Ones{Int}(∞)) # the golden ratio
ℵ₀-element PTY.ContFrac.ContinuedFraction{Rational{Int64}, Ones{Rational{Int64}, 1, Tuple{InfiniteArrays.OneToInf{Int64}}}, Ones{Rational{Int64}, 1, Tuple{InfiniteArrays.OneToInf{Int64}}}} with indices OneToInf():
   2//1
   3//2
   5//3
   8//5
  13//8
  21//13
  34//21
  55//34
  89//55
 144//89
    ⋮

PTY.ContFrac.cfrac — Method

cfrac(f, x)

The continued fraction converging to f(x). Currently supported functions: exp, log, sin, tan, csc, atanh

PTY.Infs for infinities

This is an experiment of rewriting Infinities.jl. This module is not expected to work at present.